MYR-2002: refactor interceptor
