+++ src/interceptors/index.ts
@@ -2,9 +2,6 @@ export * from './authentication.interceptor';
 export * from './authorize.interceptor';
 export * from './create.interceptor';
 export * from './delete.interceptor';
-export * from './experience.interceptor';
 export * from './find-by-id.interceptor';
 export * from './pagination.interceptor';
-export * from './report.interceptor';
 export * from './update.interceptor';
-export * from './validate-post-import-url.interceptor';
